Rename `lib` to `Library` This should be reorganized later.
