+++ kernel/Bootloader/Handoff.cpp
@@ -12,8 +12,8 @@
 #include <Bootloader/Arguments.hpp>
 #include <Bootloader/Handoff.hpp>
 // System library functions
-#include <lib/stdio.hpp>
-#include <lib/string.hpp>
+#include <Library/stdio.hpp>
+#include <Library/string.hpp>
 #include <mem/paging.hpp>
 #include <sys/Panic.hpp>
 // Generic devices
