+++ kernel/Library/mutex.cpp
@@ -18,8 +18,8 @@
 // Reference:
 // https://gcc.gnu.org/onlinedocs/gcc-8.3.0/gcc/_005f_005fatomic-Builtins.html
 
-#include <lib/mutex.hpp>
-#include <lib/errno.hpp>
+#include <Library/mutex.hpp>
+#include <Library/errno.hpp>
 
 Mutex::Mutex(const char* name)
     : locked(false)
