+++ kernel/Library/rand.hpp
