+++ kernel/Library/string.hpp
